// overlay/src/client/message_schema.ts
export class IlpFeatureExtractorDto {
    public ilp_scale: number;
    public axis_2d: 'x' | 'y' | 'z' | undefined;
    public class_name: 'Gaussian Smoothing' | 'Laplacian of Gaussian' | 'Gaussian Gradient Magnitude' | 'Difference of Gaussians' | 'Structure Tensor Eigenvalues' | 'Hessian of Gaussian Eigenvalues';
    constructor(params: {
        ilp_scale: number,
        axis_2d: 'x' | 'y' | 'z' | undefined,
        class_name: 'Gaussian Smoothing' | 'Laplacian of Gaussian' | 'Gaussian Gradient Magnitude' | 'Difference of Gaussians' | 'Structure Tensor Eigenvalues' | 'Hessian of Gaussian Eigenvalues',
    }) {
        this.ilp_scale = params.ilp_scale;
        this.axis_2d = params.axis_2d;
        this.class_name = params.class_name;
    }
    public toJsonValue(): JsonObject{
        return {
            ilp_scale: this.ilp_scale,
            axis_2d: toJsonValue(this.axis_2d),
            class_name: this.class_name,
        }
    }
    public static fromJsonValue(value: JsonValue): IlpFeatureExtractorDto | Error{
        return parse_as_IlpFeatureExtractorDto(value)    }
}
// ...
export class FeatureSelectionAppletStateDto {
    public feature_extractors: Array<IlpFeatureExtractorDto>;
    constructor(params: {
        feature_extractors: Array<IlpFeatureExtractorDto>,
    }) {
        this.feature_extractors = params.feature_extractors;
    }
    public toJsonValue(): JsonObject{
        return {
            feature_extractors: this.feature_extractors.map(item => item.toJsonValue()),
        }
    }
    public static fromJsonValue(value: JsonValue): FeatureSelectionAppletStateDto | Error{
        return parse_as_FeatureSelectionAppletStateDto(value)    }
}
// ...
export class AddFeatureExtractorsParamsDto {
    public feature_extractors: Array<IlpFeatureExtractorDto>;
    constructor(params: {
        feature_extractors: Array<IlpFeatureExtractorDto>,
    }) {
        this.feature_extractors = params.feature_extractors;
    }
    public toJsonValue(): JsonObject{
        return {
            feature_extractors: this.feature_extractors.map(item => item.toJsonValue()),
        }
    }
    public static fromJsonValue(value: JsonValue): AddFeatureExtractorsParamsDto | Error{
        return parse_as_AddFeatureExtractorsParamsDto(value)    }
}
// ...
export class RemoveFeatureExtractorsParamsDto {
    public feature_extractors: Array<IlpFeatureExtractorDto>;
    constructor(params: {
        feature_extractors: Array<IlpFeatureExtractorDto>,
    }) {
        this.feature_extractors = params.feature_extractors;
    }
    public toJsonValue(): JsonObject{
        return {
            feature_extractors: this.feature_extractors.map(item => item.toJsonValue()),
        }
    }
    public static fromJsonValue(value: JsonValue): RemoveFeatureExtractorsParamsDto | Error{
        return parse_as_RemoveFeatureExtractorsParamsDto(value)    }
}